uno_backend/requirements.txt
fastapi
uvicorn
httpx
websockets
python-dotenv